fix(utils.py): handle litellm proxy case for checking model info
