+++ litellm/utils.py
@@ -4267,6 +4267,10 @@ def _check_provider_match(model_info: dict, custom_llm_provider: Optional[str]) 
             "litellm_provider"
         ].startswith("bedrock"):
             return True
+        elif (
+            custom_llm_provider == "litellm_proxy"
+        ):  # litellm_proxy is a special case, it's not a provider, it's a proxy for the provider
+            return True
         else:
             return False
 
@@ -4319,20 +4323,6 @@ def _get_potential_model_names(
             custom_llm_provider,
             stripped_model_name,
         )
-
-    # we need to strip the litellm_proxy/ prefix to get the underlying model
-    if custom_llm_provider == "litellm_proxy":
-        # If model starts with litellm_proxy/, strip it
-        if model.startswith("litellm_proxy/"):
-            underlying_model = model.split("/", 1)[1]
-        else:
-            underlying_model = model
-        
-        # Use the underlying model for all lookups
-        split_model = underlying_model
-        stripped_model_name = underlying_model
-        combined_stripped_model_name = underlying_model
-
 
     return PotentialModelNamesAndCustomLLMProvider(
         split_model=split_model,
@@ -4407,6 +4397,7 @@ def _is_potential_model_name_in_model_cost(
         potential_model_name in litellm.model_cost
         for potential_model_name in potential_model_names.values()
     )
+
 
 def _get_model_info_helper(  # noqa: PLR0915
     model: str, custom_llm_provider: Optional[str] = None
@@ -4518,7 +4509,7 @@ def _get_model_info_helper(  # noqa: PLR0915
                 if not _check_provider_match(
                     model_info=_model_info, custom_llm_provider=custom_llm_provider
                 ):
-                    _model_info = None 
+                    _model_info = None
 
             if _model_info is None or key is None:
                 raise ValueError(
